# Don't use Vector<byte>.Dot