23% improvement in RPS

Faster implementation than using `Vector<byte>.Dot`

Resolves #493

Before
> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     8.68ms   14.73ms 333.56ms   95.18%
    Req/Sec    47.81k     4.58k  122.10k    90.28%
  44317883 requests in 30.10s, 5.45GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 1472388.96

After
> Running 30s test @ http://.../plaintext
  32 threads and 1024 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    11.18ms   28.91ms 558.18ms   94.56%
    Req/Sec    59.52k     8.66k  282.83k    93.10%
  54796839 requests in 30.10s, 6.74GB read
  Socket errors: connect 35, read 0, write 0, timeout 0
Requests/sec: 1820474.06