@benaadams 2015-12-16
Has a bug - adding more tests