@halter73 2015-12-16
23%! That's great. We're going to run this in our perflab along with the StringValues change to a class (https://github.com/aspnet/Common/pull/68) and report back with the combined results.