@darcythomas 2015-12-17
What is the environment that you ran these tests on? (So I can make some apple to apple comparisons)