@halter73 2015-12-17
1.0.0-rc2-16312 coreclr x64 win dnx on my laptop (intel core i7-3367U)