@halter73 2015-12-17
It's hard to make apple-to-apple comparisons unless you run the baseline on the exact same machine.