@darcythomas 2015-12-17
Apple to pear comparisons perhaps :P 

I was wondering if it was something "Standard-ish" like a A3 VM instance on Azure, or something...