@benaadams 2015-12-17
@darcythomas you need to run in release (SIMD optimizations are weakened in debug; so its possible to debug them). `Vector<float>.Dot` is hardware accelerated, but there isn't an equivalent cpu instruction for `Vector<byte>.Dot`.

On a A3 VM you might max out its network first; I'm testing with G4 VMs as already above 2.3GBps sends.